refactor(telemetry): decompose metrics.go into 4 files (<500 LOC each)
Split the 919-LOC metrics.go into:
- metrics.go (265 LOC): registration, standalone helpers, core types
- metrics_tracker.go (227 LOC): sessionCostTracker and all its methods
- metrics_cost.go (230 LOC): EstimateCost, ledger recording, retention
- metrics_summary.go (247 LOC): getCostSummary, aggregation, formatting

Pure structural change — no semantic modifications. Coverage unchanged
at 84.0%, all tests pass with -race, go vet clean, verify_architecture
shows zero new edges.
